Add dummy supoort mankey patchig on startup #PAMP-3 Seems2be
